+++ stereo_matching/kernels/oii_hcross.cl
@@ -3,8 +3,8 @@
 __kernel void Oii_hcross (
 	__global int* cross_l,
  __global int* cross_r,
- __global int* cost,
- __global int* temp_cost,
+ __global float* cost,
+ __global float* temp_cost,
  __global int* size
  )
 {
@@ -26,7 +26,7 @@ __kernel void Oii_hcross (
  int h_plus = min(h_plus_r, h_plus_l);
 
  int delta = h_plus - h_minus;
- int new_one = (int)((cost[min(size[0]-1, pos.x + h_plus) + size[0] * pos.y + area * pos.z] - cost[max(0, pos.x + h_minus - 1) + size[0] * pos.y + area * pos.z]) / delta);
+ float new_one = (cost[min(size[0]-1, pos.x + h_plus) + size[0] * pos.y + area * pos.z] - cost[max(0, pos.x + h_minus - 1) + size[0] * pos.y + area * pos.z]) / delta;
 // barrier(CLK_LOCAL_MEM_FENCE | CLK_GLOBAL_MEM_FENCE);
  temp_cost[pos.x + size[0] * pos.y + area * pos.z] = new_one;
 }
